Debug messages from the 4S!

+++ carbon/atmegau2/debug.c
@@ -4,7 +4,7 @@
 volatile uint8_t debug_flush_timer=0;
 
 // transmit a character.  0 returned on success, -1 on error
-int8_t usb_debug_putchar(uint8_t c)
+int usb_debug_putchar(uint8_t c)
 {
 	static uint8_t previous_timeout=0;
 	uint8_t timeout, intr_state;
